Spawn PCBaseTypeID to enum. Store the control tower's ID on the PC base for quicker retrieval. Use NWN math functions due to accuracy differences between platforms.

+++ SWLOR.Game.Server/Conversation/ApartmentRental.cs
@@ -270,7 +270,7 @@ namespace SWLOR.Game.Server.Conversation
             {
                 PlayerID = player.GlobalID,
                 BuildingStyleID = data.BuildingStyleID,
-                PCBaseTypeID = (int)PCBaseType.Apartment,
+                PCBaseTypeID = PCBaseType.Apartment,
                 ApartmentBuildingID = data.ApartmentType,
                 CustomName = string.Empty,
                 DateInitialPurchase = DateTime.UtcNow,
